Refactor: Centralize time validation logic in TimeValidationHelper
This commit refactors the `ValidationEditHelper` module to delegate
time-specific validation and manipulation logic to the newly included
`TimeValidationHelper` module.

Key changes include:
- Moved methods for parsing time strings, adjusting end datetimes,
  determining base date/time, creating new datetimes, and validating
  future dates, start times, and end times to `TimeValidationHelper`.
- Updated `ValidationEditHelper` to call these methods from the
  `TimeValidationHelper` module.
- Refactored `perform_validation` to accept an options hash and made it
  a private method.
- The `determine_and_create_datetime` method now calls the corresponding
  methods from `TimeValidationHelper`.
- Changed visibility of `validate_collision` to `protected`.

This refactoring improves code organization by centralizing
time-related validation logic, making the `ValidationEditHelper`
cleaner and more focused on its primary responsibility of orchestrating
the validation and update process for item fields.
